Create translate.py
Uses tesseract to take "input_image.jpg" with one line of japanese text in it and then send it to Sugoi Translator to translate into english and then attempts to find the center point of the text box and place the text there. (That ended up being more complicated than I thought. I'm admittedly not very good with python but the fact that it even works is a miracle at this point